+++ packages/dumb/src/utils/datetime/datetime.util.ts
@@ -35,6 +35,8 @@ const t = i18next.t.bind(i18next)
 export const ONE_HOUR_MS = 3600000
 export const HOURS_IN_DAY = 24
 export const TIMEZONE_UTC = 'UTC'
+export const THIRTY_MINS = ONE_HOUR_MS / 2
+export const TWENTY_FOUR_HRS = ONE_HOUR_MS * 24
 
 /**
  * getHourMinuteFormat