+++ packages/viz/src/components/trends/common/TrendsSVGContainer.js
@@ -32,10 +32,9 @@ import { MGDL_UNITS, MMOLL_UNITS, TimeService } from 'medical-domain'
 import { THREE_HRS } from '../../../utils/datetime'
 import { findDatesIntersectingWithCbgSliceSegment } from '../../../utils/trends/data'
 import CBGDateTracesAnimationContainer from '../cbg/CBGDateTracesAnimationContainer'
-import CBGSlicesContainer from '../cbg/CBGSlicesContainer'
 import FocusedCBGSliceSegment from '../cbg/FocusedCBGSliceSegment'
 import TargetRangeLines from './TargetRangeLines'
-import { Background, NoDataLabel, XAxisLabels, XAxisTicks, YAxisLabelsAndTicks } from 'dumb'
+import { Background, CbgSlicesContainer, NoDataLabel, XAxisLabels, XAxisTicks, YAxisLabelsAndTicks } from 'dumb'
 
 const BUMPERS = {
   top: 50,
@@ -167,7 +166,7 @@ export class TrendsSVGContainer extends React.Component {
 
     return (
       <g id="cbgTrends">
-        <CBGSlicesContainer
+        <CbgSlicesContainer
           bgBounds={this.props.bgPrefs.bgBounds}
           sliceWidth={sliceWidth}
           data={this.props.cbgData}